requirements/requirements.txt
pyyaml
omegaconf
torch
torchvision
tensorboard
matplotlib
tqdm
opencv-python